feat: 컴포넌트 세부 분리, 반응형 리퀘스트 작업
