+++ src/app/community/_components/BestArticles/BestArticlesLoad.jsx
@@ -1,70 +1,10 @@
-"use client";
-
-import { getArticles } from "@/lib/api/article.api";
 import Image from "next/image";
-import React, { useEffect, useState } from "react";
 import ic_heart from "@/assets/images/common/ic_heart.svg";
 import img_badge from "@/assets/images/community/img_badge.svg";
 import img_notebook from "@/assets/images/community/img_notebook.svg";
 import Link from "next/link";
-import { useGetDeviceType } from "@/hooks/useGetDeviceType";
 
-export default function BestArticleList() {
-  const [params, setParams] = useState({
-    page: 1,
-    limit: 5,
-    orderBy: "recent",
-    keyword: "",
-  });
-
-  const bestArticles = [
-    {
-      id: "1",
-      title: "맥북 16인치 16기가 1테라 정도 사양이면 얼마에 팔아야 하나요?",
-      createdAt: "2025. 04. 16",
-    },
-    {
-      id: "2",
-      title: "맥북 16인치 16기가 1테라 정도 사양이면 얼마에 팔아야 하나요?",
-      createdAt: "2025. 04. 16",
-    },
-    {
-      id: "3",
-      title: "맥북 16인치 16기가 1테라 정도 사양이면 얼마에 팔아야 하나요?",
-      createdAt: "2025. 04. 16",
-    },
-  ];
-
-  // const [articles, setArticles] = useState([]);
-  // const [isLoading, setIsLoading] = useState(true);
-
-  // 반응형 리퀘스트 보내기
-  useGetDeviceType(setParams);
-
-  // useEffect(() => {
-  //   articlesLoad(params);
-  // }, [params]);
-
-  // const articlesLoad = async (params) => {
-  //   const articles = await getArticles(params);
-  //   setArticles(articles);
-  //   setIsLoading(false);
-  // };
-
-  return (
-    <div className="flex flex-col gap-[16px]">
-      <h2 className="font-bold text-[20px]">베스트 게시글</h2>
-      <div className="grid grid-cols-1 gap-[16px] sm:grid-cols-2 md:grid-cols-3 md:gap-[24px]">
-        {/* {isLoading ? "게시글 로딩중..." : }*/}
-        {bestArticles.map((bestArticle) => (
-          <BestAriclesLoad key={bestArticle.id} bestArticle={bestArticle} />
-        ))}
-      </div>
-    </div>
-  );
-}
-
-function BestAriclesLoad({ bestArticle }) {
+export default function BestArticlesLoad({ bestArticle }) {
   return (
     <Link
       href={`/community/${bestArticle.id}`}
